+++ doc/api/fs.md
@@ -3382,9 +3382,9 @@ Synchronous versions of [`fs.write()`][]. Returns the number of bytes written.
 
 > Stability: 1 - Experimental
 
-The `fs/promises` API provides an alternative set of asynchronous file system
+The `fs.promises` API provides an alternative set of asynchronous file system
 methods that return `Promise` objects rather than using callbacks. The
-API is accessible via `require('fs/promises')`.
+API is accessible via `require('fs').promises`.
 
 ### class: FileHandle
 <!-- YAML
