+++ lib/fs.js
@@ -45,7 +45,8 @@ const { Readable, Writable } = require('stream');
 const EventEmitter = require('events');
 const { FSReqWrap, statValues, kFsStatsFieldsLength } = binding;
 const { FSEvent } = process.binding('fs_event_wrap');
-const internalFS = require('internal/fs');
+const promises = require('internal/fs/promises');
+const internalFS = require('internal/fs/utils');
 const { getPathFromURL } = require('internal/url');
 const internalUtil = require('internal/util');
 const {
@@ -71,6 +72,21 @@ const {
   CHAR_FORWARD_SLASH,
   CHAR_BACKWARD_SLASH,
 } = require('internal/constants');
+
+let warn = true;
+
+Object.defineProperty(fs, 'promises', {
+  configurable: true,
+  enumerable: true,
+  get() {
+    if (warn) {
+      warn = false;
+      process.emitWarning('The fs.promises API is experimental',
+                          'ExperimentalWarning');
+    }
+    return promises;
+  }
+});
 
 Object.defineProperty(exports, 'constants', {
   configurable: false,
