+++ lib/internal/fs/promises.js
@@ -1,8 +1,5 @@
 'use strict';
 
-process.emitWarning('The fs/promises API is experimental',
-                    'ExperimentalWarning');
-
 const {
   F_OK,
   O_SYMLINK,
@@ -37,7 +34,7 @@ const {
   validateOffsetLengthWrite,
   validatePath,
   validateUint32
-} = require('internal/fs');
+} = require('internal/fs/utils');
 const pathModule = require('path');
 
 const kHandle = Symbol('handle');
