+++ lib/internal/fs/utils.js
