+++ lib/internal/modules/cjs/loader.js
@@ -28,7 +28,7 @@ const { getURLFromFilePath } = require('internal/url');
 const vm = require('vm');
 const assert = require('assert').ok;
 const fs = require('fs');
-const internalFS = require('internal/fs');
+const internalFS = require('internal/fs/utils');
 const path = require('path');
 const {
   internalModuleReadJSON,
