+++ lib/internal/modules/esm/default_resolve.js
@@ -2,7 +2,7 @@
 
 const { URL } = require('url');
 const CJSmodule = require('internal/modules/cjs/loader');
-const internalFS = require('internal/fs');
+const internalFS = require('internal/fs/utils');
 const { NativeModule, internalBinding } = require('internal/bootstrap/loaders');
 const { extname } = require('path');
 const { realpathSync } = require('fs');
