+++ lib/internal/process/stdio.js
@@ -158,7 +158,7 @@ function createWritableStdioStream(fd) {
       break;
 
     case 'FILE':
-      var fs = require('internal/fs');
+      var fs = require('internal/fs/utils');
       stream = new fs.SyncWriteStream(fd, { autoClose: false });
       stream._type = 'fs';
       break;
