+++ node.gyp
@@ -39,7 +39,6 @@
       'lib/domain.js',
       'lib/events.js',
       'lib/fs.js',
-      'lib/fs/promises.js',
       'lib/http.js',
       'lib/http2.js',
       'lib/_http_agent.js',
@@ -103,7 +102,8 @@
       'lib/internal/errors.js',
       'lib/internal/fixed_queue.js',
       'lib/internal/freelist.js',
-      'lib/internal/fs.js',
+      'lib/internal/fs/promises.js',
+      'lib/internal/fs/utils.js',
       'lib/internal/http.js',
       'lib/internal/inspector_async_hook.js',
       'lib/internal/linkedlist.js',
