+++ test/parallel/test-fs-promises-file-handle-append-file.js
@@ -2,11 +2,11 @@
 
 const common = require('../common');
 
-// The following tests validate base functionality for the fs/promises
+// The following tests validate base functionality for the fs.promises
 // FileHandle.appendFile method.
 
 const fs = require('fs');
-const { open } = require('fs/promises');
+const { open } = fs.promises;
 const path = require('path');
 const tmpdir = require('../common/tmpdir');
 const assert = require('assert');
