+++ test/parallel/test-fs-promises-file-handle-stat.js
@@ -2,10 +2,10 @@
 
 const common = require('../common');
 
-// The following tests validate base functionality for the fs/promises
+// The following tests validate base functionality for the fs.promises
 // FileHandle.stat method.
 
-const { open } = require('fs/promises');
+const { open } = require('fs').promises;
 const path = require('path');
 const tmpdir = require('../common/tmpdir');
 const assert = require('assert');
