+++ test/parallel/test-fs-promises-writefile.js
@@ -2,7 +2,7 @@
 
 const common = require('../common');
 const fs = require('fs');
-const fsPromises = require('fs/promises');
+const fsPromises = fs.promises;
 const path = require('path');
 const tmpdir = require('../common/tmpdir');
 const assert = require('assert');
