+++ test/parallel/test-fs-syncwritestream.js
@@ -6,8 +6,8 @@ const stream = require('stream');
 const fs = require('fs');
 const path = require('path');
 
-// require('internal/fs').SyncWriteStream is used as a stdio implementation
-// when stdout/stderr point to files.
+// require('internal/fs/utils').SyncWriteStream is used as a stdio
+// implementation when stdout/stderr point to files.
 
 if (process.argv[2] === 'child') {
   // Note: Calling console.log() is part of this test as it exercises the
