+++ test/sequential/test-async-wrap-getasyncid.js
@@ -3,7 +3,7 @@
 const common = require('../common');
 const assert = require('assert');
 const fs = require('fs');
-const fsPromises = require('fs/promises');
+const fsPromises = fs.promises;
 const net = require('net');
 const providers = Object.assign({}, process.binding('async_wrap').Providers);
 const fixtures = require('../common/fixtures');
